+++ doc/changelog.rst
@@ -3,7 +3,7 @@ Changelog
 
 v0.3.1
 ------
-*(not-released-yet)*
+*(2017-11-21)*
 
 - Fixed exception with evaluation of grouped categorical columns when
   there are missing categories in the data.